--- account_bank_statement_import_camt/i18n/es.po
@@ -1,55 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# * account_bank_statement_import_camt
-# 
-# Translators:
-# OCA Transbot <transbot@odoo-community.org>, 2017
-# enjolras <yo@miguelrevilla.com>, 2018
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 10.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2018-02-21 01:41+0000\n"
-"PO-Revision-Date: 2018-02-21 01:41+0000\n"
-"Last-Translator: enjolras <yo@miguelrevilla.com>, 2018\n"
-"Language-Team: Spanish (https://www.transifex.com/oca/teams/23907/es/)\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Language: es\n"
-"Plural-Forms: nplurals=2; plural=(n != 1);\n"
-
-#. module: account_bank_statement_import_camt
-#: model:ir.ui.view,arch_db:account_bank_statement_import_camt.account_bank_statement_import_view
-msgid "CAMT"
-msgstr "CAMT"
-
-#. module: account_bank_statement_import_camt
-#: model:ir.model.fields,field_description:account_bank_statement_import_camt.field_account_bank_statement_import_camt_parser_display_name
-msgid "Display Name"
-msgstr "Nombre a mostrar"
-
-#. module: account_bank_statement_import_camt
-#: model:ir.model.fields,field_description:account_bank_statement_import_camt.field_account_bank_statement_import_camt_parser_id
-msgid "ID"
-msgstr "ID"
-
-#. module: account_bank_statement_import_camt
-#: model:ir.model,name:account_bank_statement_import_camt.model_account_bank_statement_import
-msgid "Import Bank Statement"
-msgstr "Importar extracto bancario"
-
-#. module: account_bank_statement_import_camt
-#: model:ir.model.fields,field_description:account_bank_statement_import_camt.field_account_bank_statement_import_camt_parser___last_update
-msgid "Last Modified on"
-msgstr "Última modificación en"
-
-#. module: account_bank_statement_import_camt
-#: model:ir.model,name:account_bank_statement_import_camt.model_account_bank_statement_import_camt_parser
-msgid "account.bank.statement.import.camt.parser"
-msgstr "account.bank.statement.import.camt.parser"
-
-#. module: account_bank_statement_import_camt
-#: model:ir.ui.view,arch_db:account_bank_statement_import_camt.account_bank_statement_import_view
-msgid "zipped CAMT"
-msgstr "CAMT en .zip"
