--- account_bank_statement_import_camt/i18n/hr.po
@@ -1,54 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# * account_bank_statement_import_camt
-# 
-# Translators:
-# Bole <bole@dajmi5.com>, 2018
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 10.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2018-02-21 01:41+0000\n"
-"PO-Revision-Date: 2018-02-21 01:41+0000\n"
-"Last-Translator: Bole <bole@dajmi5.com>, 2018\n"
-"Language-Team: Croatian (https://www.transifex.com/oca/teams/23907/hr/)\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Language: hr\n"
-"Plural-Forms: nplurals=3; plural=n%10==1 && n%100!=11 ? 0 : n%10>=2 && n%10<=4 && (n%100<10 || n%100>=20) ? 1 : 2;\n"
-
-#. module: account_bank_statement_import_camt
-#: model:ir.ui.view,arch_db:account_bank_statement_import_camt.account_bank_statement_import_view
-msgid "CAMT"
-msgstr "CAMT"
-
-#. module: account_bank_statement_import_camt
-#: model:ir.model.fields,field_description:account_bank_statement_import_camt.field_account_bank_statement_import_camt_parser_display_name
-msgid "Display Name"
-msgstr "Naziv"
-
-#. module: account_bank_statement_import_camt
-#: model:ir.model.fields,field_description:account_bank_statement_import_camt.field_account_bank_statement_import_camt_parser_id
-msgid "ID"
-msgstr "ID"
-
-#. module: account_bank_statement_import_camt
-#: model:ir.model,name:account_bank_statement_import_camt.model_account_bank_statement_import
-msgid "Import Bank Statement"
-msgstr "Uvoz bankovnog izvoda"
-
-#. module: account_bank_statement_import_camt
-#: model:ir.model.fields,field_description:account_bank_statement_import_camt.field_account_bank_statement_import_camt_parser___last_update
-msgid "Last Modified on"
-msgstr "Zadnje modificirano"
-
-#. module: account_bank_statement_import_camt
-#: model:ir.model,name:account_bank_statement_import_camt.model_account_bank_statement_import_camt_parser
-msgid "account.bank.statement.import.camt.parser"
-msgstr "account.bank.statement.import.camt.parser"
-
-#. module: account_bank_statement_import_camt
-#: model:ir.ui.view,arch_db:account_bank_statement_import_camt.account_bank_statement_import_view
-msgid "zipped CAMT"
-msgstr "kompresirani CAMT"
