+++ account_bank_statement_import_camt_oca/test_files/golden-camt053.pydata
