+++ account_bank_statement_import_camt_oca/test_files/test-camt053-txdtls
